backend/requirements.txt
## The following requirements were added by pip freeze:
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.115.13
h11==0.16.0
idna==3.10
joblib==1.5.1
numpy==2.2.6
pillow==11.2.1
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
scikit-learn==1.7.0
scipy==1.15.3
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3